Alterações
tudo que estava como float alterei para double

+++ Financas/Models/ContasPagar.cs
@@ -9,7 +9,7 @@ namespace Models
     public class ContasPagar
     {
         public int Id { get; set; }
-        public float ValorPagar { get; set; }
+        public double ValorPagar { get; set; }
         public string Descricao { get; set; }
     }
 }
